lib+workspace+http: Remove `lib.BlockBuf`
We tried to use `lib.BlockBuf` to reduce the number of allocations by
passing it to `lib.Storage.ReadBlock`, `lib.Storage.WriteBlock`, etc.

First, the implementation was fundamentally wrong - we were passing
the `lib.BlockBuf` (an array-type) by value, which means it has been
copied all over the place.

We did try to fix this by passing `*lib.BlockBuf` but that made the code
more fragile.

So, we removed `lib.BlockBuf` and create byte slices where needed.

Next time we try to optimize something like that, we should probably
test its effects beforehand. :-)